+++ R/io_writeLAS.R
@@ -34,6 +34,7 @@
 #'
 #' @param las an object of class LAS.
 #' @param file character. A character string naming an output file.
+#' @param index boolean. Also write a lax file to index the points in the files
 #'
 #' @return Nothing. This function is used for its side-effect of writing a file.
 #'
@@ -44,10 +45,11 @@
 #' las = readLAS(LASfile)
 #' subset = clip_rectangle(las, 684850, 5017850, 684900, 5017900)
 #' writeLAS(subset, tempfile(fileext = ".laz"))
-writeLAS = function(las, file)
+writeLAS = function(las, file, index = FALSE)
 {
   stopifnotlas(las)
   assert_is_a_string(file)
+  assert_is_a_bool(index)
 
   if (is.empty(las)) stop("Cannot write a file with 0 point", call. = FALSE)
 
@@ -57,5 +59,8 @@ writeLAS = function(las, file)
   if (!islas) stop(glue::glue("File(s) {file} not supported"), call. = FALSE)
 
   rlas::write.las(file, as.list(las@header), las@data)
-  return(invisible())
+
+  if (index) rlas:::writelax(file)
+
+  return(invisible(file))
 }